Refactor: Extract shared comment display utilities
Move duplicated formatRelativeTime, getAuthorDisplay, countComments,
and printCommentTree functions into src/lib/comments-display.ts.
